+++ go/preflight_kit_sdk/preflight_sdk_integration_test.go
@@ -128,15 +128,17 @@ func testcaseHeartbeatTimeout(t *testing.T, op PreflightOperations) {
 
 func testCaseStatusWithGenericError(t *testing.T, op PreflightOperations) {
 	op.preflight.statusError = fmt.Errorf("this is a test error")
-	_, err := op.statusResult(t, preflight_kit_api.PreflightState{})
-	assert.Equal(t, &preflight_kit_api.PreflightKitError{Title: "Failed to read status.", Detail: extutil.Ptr("this is a test error")}, err)
+	statusResult, err := op.statusResult(t, preflight_kit_api.PreflightState{})
+	assert.Equal(t, &preflight_kit_api.PreflightKitError{Title: "Failed to read preflight status.", Detail: extutil.Ptr("this is a test error")}, statusResult.Error)
+	assert.Nil(t, err)
 	op.assertCall(t, "Status", ANY_ARG)
 }
 
 func testCaseStatusWithExtensionKitError(t *testing.T, op PreflightOperations) {
-	op.preflight.statusError = extutil.Ptr(extension_kit.ToError("this is a test error", errors.New("with some setails")))
-	_, response := op.statusResult(t, preflight_kit_api.PreflightState{})
-	assert.Equal(t, &preflight_kit_api.PreflightKitError{Title: "this is a test error", Detail: extutil.Ptr("with some setails")}, response)
+	op.preflight.statusError = extutil.Ptr(extension_kit.ToError("this is a test error", errors.New("with some details")))
+	statusResult, err := op.statusResult(t, preflight_kit_api.PreflightState{})
+	assert.Equal(t, &preflight_kit_api.PreflightKitError{Title: "this is a test error", Detail: extutil.Ptr("with some details")}, statusResult.Error)
+	assert.Nil(t, err)
 	op.assertCall(t, "Status", ANY_ARG)
 }